requirements.txt
fastapi
uvicorn[standard]
langchain
langchain_community
langchain-huggingface
langchain_openai
PyPDF2
tiktoken
python-dotenv
sympy
wikipedia
faiss-cpu
transformers
sentence-transformers